002: Bugfix Release April 5th, 2005
Version 002 fixes all known bugs. It also adds support for curses
output. This allows one to play the game over a text only terminal. I
have hopes it also makes it easy to port to the Mac, but I have not yet
got a Mac to try it on. If you port it to OSX, I would be grateful if
you could send me the binary so I can release it to others.